Rename LabelsOBB -> OBBBuffer
