+++ core/src/labels/labels.cpp
@@ -13,6 +13,7 @@
 #include "labels/labelSet.h"
 #include "labels/textLabel.h"
 #include "marker/marker.h"
+#include "labels/obbBuffer.h"
 
 #include "glm/glm.hpp"
 #include "glm/gtc/matrix_transform.hpp"
@@ -324,7 +325,7 @@ void Labels::handleOcclusions(const ViewState& _viewState) {
         }
 
         ScreenTransform transform { m_transforms, entry.transform };
-        LabelOBBs obbs { m_obbs, entry.obbs, true };
+        OBBBuffer obbs { m_obbs, entry.obbs, true };
 
         l->obbs(transform, obbs);
 
@@ -439,7 +440,7 @@ void Labels::updateLabelSet(const ViewState& _viewState, float _dt,
         m_needUpdate |= entry.label->evalState(_dt);
 
         if (entry.label->visibleState()) {
-            for (auto& obb : LabelOBBs{ m_obbs, entry.obbs }) {
+            for (auto& obb : OBBBuffer{ m_obbs, entry.obbs }) {
 
                 if (obb.getExtent().intersect(screenBounds)) {
                     entry.label->addVerticesToMesh(transform, _viewState.viewportSize);
@@ -500,7 +501,7 @@ void Labels::drawDebug(RenderState& rs, const View& _view) {
         }
 #endif
 
-        for (auto& obb : LabelOBBs{ m_obbs, entry.obbs }) {
+        for (auto& obb : OBBBuffer{ m_obbs, entry.obbs }) {
             Primitives::drawPoly(rs, &(obb.getQuad())[0], 4);
         }
